Revert "Feat/20230710 more improvements"
