+++ Cargo.toml
@@ -24,4 +24,4 @@ clap = { version = "4.2.4", features = ["derive"] }
 anyhow = "*"
 reqwest = { version = "0.11", features = ["json"] }
 tokio = { version = "1", features = ["full"] }
-ethers = "2.0"
+
