+++ src/main.rs
@@ -2,10 +2,11 @@
 use circuit_definitions::circuit_definitions::recursion_layer::scheduler::ConcreteSchedulerCircuitBuilder;
 use clap::Parser;
 use colored::Colorize;
+use serde::Deserialize;
+use std::fs;
 use std::fs::File;
 use std::io::Read;
-
-mod requests;
+use std::io::Cursor;
 
 use boojum::{
     cs::implementations::{
@@ -28,22 +29,18 @@ pub enum FriProofWrapper {
 #[command(author = "Matter Labs", version, about = "Boojum CLI verifier", long_about = None)]
 struct Cli {
     #[arg(long)]
-    // Path to the .bin file with the proof
+    /// Path to the .bin file with the proof
     proof: Option<String>,
     #[arg(long)]
-    // Batch number to check proof for
+    /// Batch number to check proof for
     batch: Option<usize>,
     #[arg(long, default_value = "mainnet")]
-    // Batch number to check proof for
+    /// Batch number to check proof for
     network: String,
-    #[arg(long)]
-    // RPC endpoint to use to fetch L1 information
-    l1_rpc: Option<String>,
 }
 
-
 /// Reads proof (in FriProofWrapper format) from a given bin file.
-pub fn proof_from_file<T: for<'a> serde::Deserialize<'a>>(proof_path: &str) -> T {
+pub fn proof_from_file<T: for<'a> Deserialize<'a>>(proof_path: &str) -> T {
     let mut file = File::open(proof_path).unwrap();
     let mut buffer = Vec::new();
     file.read_to_end(&mut buffer).unwrap();
@@ -80,6 +77,30 @@ pub fn verify_scheduler_proof(proof_path: &str) -> anyhow::Result<String> {
     }
 }
 
+/// Download the proof file if it exists and saves locally
+async fn fetch_proof_from_storage(batch_number: usize, network: String) -> Result<String, Box<dyn std::error::Error>> {
+
+    println!("Downloading proof for batch {} on network {}", batch_number, network);
+
+    let client = reqwest::Client::new();
+    let url = format!("https://storage.googleapis.com/zksync-era-{}-proofs/proofs_fri/proof_{}.bin", network, batch_number);
+    let proof = client.get(url).send()
+        .await?;
+
+    if proof.status().is_success() {
+        fs::create_dir_all("./downloaded_proofs")?;
+        let file_path = format!("./downloaded_proofs/proof_{}_{}.bin", network, batch_number);
+
+        let mut file = std::fs::File::create(file_path.clone())?;
+        let mut content =  Cursor::new(proof.bytes().await?);
+        std::io::copy(&mut content, &mut file)?;
+
+        return Ok(file_path);
+    } else {
+        return Err(format!("Proof for batch {} on network {} not found", batch_number, network).into());
+    }
+}
+
 #[tokio::main]
 async fn main() {
     let opt = Cli::parse();
@@ -87,7 +108,6 @@ async fn main() {
     let batch_number = &opt.batch;
     let proof;
     let network = &opt.network;
-    let l1_rpc = &opt.l1_rpc;
 
     if network.to_string() != "testnet" && network.to_string() != "mainnet" {
         println!("Invalid network name. Please use 'testnet' or 'mainnet'");
@@ -95,31 +115,12 @@ async fn main() {
     }
 
     if !batch_number.is_none() {
-
-        let proof_response = requests::fetch_proof_from_storage(batch_number.unwrap(), network.to_string()).await;
+        let proof_response = fetch_proof_from_storage(batch_number.unwrap(), network.to_string()).await;
 
         if let Err(_err) = proof_response {
             println!("{}", _err);
             return
         }
-
-        if l1_rpc.is_none() {
-            println!("Skipping building batch information from Ethereum as no RPC url was provided");
-        } else {
-            let tx_hash = requests::fetch_batch_info(batch_number.unwrap(), network.to_string()).await;
-            if let Err(_err) = tx_hash {
-                println!("{}", _err);
-                return
-            }
-            let l1_batch_info = requests::fetch_l1_info(tx_hash.unwrap(), l1_rpc.clone().unwrap(), batch_number.unwrap(), network.to_string()).await;
-            if let Err(_err) = l1_batch_info {
-                println!("{}", _err);
-                return
-            }
-
-            //TODO: Reform block header with proof information and compare with L1 information
-        }
-       
         proof = proof_response.unwrap()
     } else {
         proof = (&opt.proof).clone().unwrap();
@@ -152,7 +153,7 @@ mod test {
     use super::*;
     #[test]
     fn test_scheduler_proof() {
-        verify_scheduler_proof("example_proofs/proof_52272951.bin").expect("FAILED");
+        verify_scheduler_proof("scheduler_proof/proof_52272951.bin").expect("FAILED");
     }
     #[test]
